pyproject.toml
[project]
name = "dict-back-end-programming-lopatin-danylo-538"
version = "0.1.0"
description = "A study project for back-end programming using FastAPI, Jinja2, MySQL Connector, and Uvicorn."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "mysql-connector-python>=9.6.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "starlette[full]>=0.52.1",
    "uvicorn>=0.42.0",
]
